Website V2
The entire site has been revamped from the ground up. We discussed changing the site a few weeks ago to ensure that the site is that of it's own original design rather than a copy of the UCLan site.
This push removes the old site in it's entirety and is replaced with the new site. List of changes include:
- Loading screen (Still being polished)
- New pages
- New images
- New videos

Note: The pages still need a lot of work and the pages as of right now are a template that needs to be filled in with the appropriate information.